+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+- Changed the condition type of `HAS` arguments in `@whereConditions` to always use `String` type
+
 ## v6.63.1
 
 ### Fixed